Fix unserialization of abstract twig descendant.
If an abstract class that extends the symfony twig class exists in the trace heirarchy the unserialize method in ExceptionCaster throws an fatal error.  In order to fix this I use reflection to check if the class is instantiable.  I thought this was preferable to a try/catch as the behavior isn't exceptional and should be checked.  Performance wise the try/catch may be faster but I haven't benchmarked it.